move mvc config into config class
