+++ src/main/java/uk/gov/companieshouse/bankruptofficersearch/api/BankruptOfficerSearchApiApplication.java
@@ -8,28 +8,11 @@ import uk.gov.companieshouse.bankruptofficersearch.api.interceptor.Authorisation
 import uk.gov.companieshouse.bankruptofficersearch.api.interceptor.LoggingInterceptor;
 
 @SpringBootApplication
-public class BankruptOfficerSearchApiApplication implements WebMvcConfigurer {
+public class BankruptOfficerSearchApiApplication {
 
     public static final String APPLICATION_NAME_SPACE = "bankrupt-officer-search-api";
 
-    private final LoggingInterceptor loggingInterceptor;
-
-    private final AuthorisationInterceptor authorisationInterceptor;
-
-    public BankruptOfficerSearchApiApplication(LoggingInterceptor loggingInterceptor, AuthorisationInterceptor authorisationInterceptor) {
-        this.loggingInterceptor = loggingInterceptor;
-        this.authorisationInterceptor = authorisationInterceptor;
-    }
-
     public static void main(String[] args) {
         SpringApplication.run(BankruptOfficerSearchApiApplication.class, args);
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(loggingInterceptor)
-                .excludePathPatterns("/internal/officer-search/scottish-bankrupt-officers/healthcheck");
-        registry.addInterceptor(authorisationInterceptor)
-                .excludePathPatterns("/internal/officer-search/scottish-bankrupt-officers/healthcheck");
     }
 }
