+++ src/test/java/uk/gov/companieshouse/bankruptofficersearch/api/BankruptOfficerSearchApiApplicationTest.java
@@ -7,8 +7,10 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.context.annotation.Import;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import uk.gov.companieshouse.bankruptofficersearch.api.config.WebMvcConfig;
 import uk.gov.companieshouse.bankruptofficersearch.api.interceptor.AuthorisationInterceptor;
 import uk.gov.companieshouse.bankruptofficersearch.api.interceptor.LoggingInterceptor;
 
@@ -18,6 +20,7 @@ import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
+@Import(WebMvcConfig.class)
 class BankruptOfficerSearchApiApplicationTest {
 
 	@Mock
@@ -33,7 +36,7 @@ class BankruptOfficerSearchApiApplicationTest {
 	private InterceptorRegistration interceptorRegistration;
 
 	@InjectMocks
-	private BankruptOfficerSearchApiApplication application;
+	private WebMvcConfig config;
 
 	@Test
 	void testAddInterceptors() {
@@ -42,7 +45,7 @@ class BankruptOfficerSearchApiApplicationTest {
 
 		when(interceptorRegistration.excludePathPatterns(anyString())).thenReturn(interceptorRegistration);
 
-		application.addInterceptors(interceptorRegistry);
+        config.addInterceptors(interceptorRegistry);
 
 		InOrder inOrder = Mockito.inOrder(interceptorRegistry);
 		inOrder.verify(interceptorRegistry).addInterceptor(loggingInterceptor);
